RAG/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "CPIC Pharmacogenomics RAG using HuggingFace embeddings + Gemini LLM"
requires-python = ">=3.11"

dependencies = [
    "langchain",
    "langchain-community",
    "langchain-text-splitters",
    "langchain-chroma",
    "langchain-huggingface",
    "chromadb",
    "sentence-transformers",
    "pypdf",
    "python-dotenv",
    "google-generativeai"
]

